Move panel placeholder and overview actions
